매입 하루 조회 [#24]

+++ src/main/java/com/haniwon/dto/outcome/response/OutcomeResponseDTO.java
@@ -19,7 +19,7 @@ public class OutcomeResponseDTO {
 
     private final String item;
     //금액
-    private final Integer amount;
+    private final Long amount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private final String memo;
